refactor issue module
