+++ src/Issue/index.js
@@ -1,39 +1,3 @@
-import React from 'react';
-import { withState } from 'recompose';
+import IssueList from './IssueList';
 
-import Button from '../Button';
-import Comments from '../Comments';
-import Link from '../Link';
-
-import './style.css';
-
-const Issue = ({
-  issue,
-  repositoryOwner,
-  repositoryName,
-  isShowComments,
-  onShowComments,
-}) => (
-  <div className="Issue">
-    <Button onClick={() => onShowComments(!isShowComments)}>
-      {isShowComments ? '-' : '+'}
-    </Button>
-
-    <div className="Issue-content">
-      <h3>
-        <Link href={issue.url}>{issue.title}</Link>
-      </h3>
-      <div dangerouslySetInnerHTML={{ __html: issue.bodyHTML }} />
-
-      {isShowComments && (
-        <Comments
-          repositoryOwner={repositoryOwner}
-          repositoryName={repositoryName}
-          issue={issue}
-        />
-      )}
-    </div>
-  </div>
-);
-
-export default withState('isShowComments', 'onShowComments', false)(Issue);
+export default IssueList;
